+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ## cordova-sqlite-evplus-feat-android-db-location 0.0.1-dev
 
+### cordova-sqlite-evcore-feature-android-db-location 0.7.2-dev1
+
+- Custom Android database file location
+
 ## cordova-plugin-sqlite-evplus-ext-common-free 0.4.1-rc01
 
 - SQLite 3.32.3 update from github:brodybits/cordova-sqlite-evplus-ext-free-dependencies#0.1.0-rc02